+++ R/reexports.R
@@ -1,0 +1,24 @@
+#' @importFrom hardhat extract_spec_parsnip
+#' @export
+hardhat::extract_spec_parsnip
+#'
+#' @importFrom hardhat extract_recipe
+#' @export
+hardhat::extract_recipe
+#'
+#' @importFrom hardhat extract_fit_parsnip
+#' @export
+hardhat::extract_fit_parsnip
+#'
+#' @importFrom hardhat extract_fit_engine
+#' @export
+hardhat::extract_fit_engine
+#'
+#' @importFrom hardhat extract_mold
+#' @export
+hardhat::extract_mold
+#'
+#' @importFrom hardhat extract_preprocessor
+#' @export
+hardhat::extract_preprocessor
+#'