+++ R/workflows-package.R
@@ -4,16 +4,14 @@
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
+#'
 #' @import rlang
-#'
-#' @importFrom generics fit
-#' @importFrom generics tidy
-#' @importFrom generics glance
 #' @importFrom generics augment
-#'
+#' @importFrom generics fit
+#' @importFrom generics glance
+#' @importFrom generics tidy
+#' @importFrom lifecycle deprecated
 #' @importFrom parsnip fit_xy
-#'
 #' @importFrom stats predict
-#'
 ## usethis namespace: end
 NULL